PDFBOX-5134: calculating the bounding boxes of the triangles reduces shading rendering times if no bounding box is preset, as suggested by Oliver Schmidtmer; Closes #109
git-svn-id: https://svn.apache.org/repos/asf/pdfbox/branches/2.0@1888288 13f79535-47bb-0310-9956-ffa450edef68
