+++ pdfbox/src/main/java/org/apache/pdfbox/pdmodel/graphics/shading/PDShadingType7.java
@@ -17,6 +17,10 @@
 package org.apache.pdfbox.pdmodel.graphics.shading;
 
 import java.awt.Paint;
+import java.awt.geom.AffineTransform;
+import java.awt.geom.Point2D;
+import java.awt.geom.Rectangle2D;
+import java.io.IOException;
 
 import org.apache.pdfbox.cos.COSDictionary;
 import org.apache.pdfbox.util.Matrix;
@@ -24,7 +28,7 @@ import org.apache.pdfbox.util.Matrix;
 /**
  * Resources for a shading type 7 (Tensor-Product Patch Mesh).
  */
-public class PDShadingType7 extends PDShadingType6
+public class PDShadingType7 extends PDMeshBasedShadingType
 {
     /**
      * Constructor using the given shading dictionary.
@@ -46,5 +50,17 @@ public class PDShadingType7 extends PDShadingType6
     public Paint toPaint(Matrix matrix)
     {
         return new Type7ShadingPaint(this, matrix);
+    }
+    
+    @Override
+    protected Patch generatePatch(Point2D[] points, float[][] color)
+    {
+        return new TensorPatch(points, color);
+    }
+
+    @Override
+    public Rectangle2D getBounds(AffineTransform xform, Matrix matrix) throws IOException
+    {
+        return getBounds(xform, matrix, 16);
     }
 }
